Updating to add padding and moving files into Feed dir.
