+++ resources/assets/components/Feed/FeedContainer.js
@@ -1,0 +1,35 @@
+import { connect } from 'react-redux';
+import Feed from './Feed';
+import { clickedViewMore, clickedSignUp } from '../../actions';
+import {
+  getBlocksWithReportbacks,
+  getVisibleBlocks,
+  getBlockOffset,
+  getMaximumOffset,
+} from '../../selectors/feed';
+
+/**
+ * Provide state from the Redux store as props for this component.
+ */
+const mapStateToProps = state => ({
+  blocks: getBlocksWithReportbacks(getVisibleBlocks(state), state),
+  canLoadMorePages: getBlockOffset(state) < getMaximumOffset(state),
+  campaignId: state.campaign.legacyCampaignId,
+  callToAction: state.campaign.callToAction,
+  signedUp: state.signups.data.includes(state.campaign.legacyCampaignId),
+  hasNewSignup: state.signups.thisSession,
+  hasPendingSignup: state.signups.isPending,
+  isAuthenticated: state.user.id !== null,
+});
+
+/**
+ * Provide pre-bound functions that allow the component to dispatch
+ * actions to the Redux store as props for this component.
+ */
+const actionCreators = {
+  clickedViewMore,
+  clickedSignUp,
+};
+
+// Export the container component.
+export default connect(mapStateToProps, actionCreators)(Feed);
